Update theme with dark mode; Note: theme is now a function useTheme()
