+++ foxglove/shared/theme/src/index.ts
@@ -1,1 +1,1 @@
-export { default } from "./theme";
+export { default } from "./useTheme";
