--- foxglove/shared/theme/src/theme.ts
@@ -1,17 +1,0 @@
-import { createTheme } from "@mui/material/styles";
-
-const theme = createTheme({
-  components: {
-    MuiTableRow: {
-      styleOverrides: {
-        root: {
-          "&:last-child td, &:last-child th": {
-            border: 0,
-          },
-        },
-      },
-    },
-  },
-});
-
-export default theme;
